+++ components/LocationFinder/index.tsx
@@ -1,0 +1,2 @@
+export {default} from './LocationFinder'
+
