+++ components/Map/index.tsx
@@ -1,1 +1,9 @@
-export {default} from './Map'
+import dynamic from 'next/dynamic'
+
+const Map = dynamic(() => import('./Map'), {
+  ssr: false
+})
+
+
+
+export default Map;
