+++ types/tankstellen-types.ts
@@ -1,33 +1,25 @@
-export type Base = {
-  status: string;
-  e5: number;
-  e10: number;
-  diesel: number;
-}
-
 export type LatLng = {
   lat: number;
   lng: number;
 };
 
-export type Location = {
-  dist: number;
-  latLng: LatLng;
 
-}
-
-
-
-export type TankstellenListe = Base & LatLng & {
-  id: string;
-  name: string;
-  brand: string;
-  street: string;
-  place: string;
-  diesel: number;
+export type Station = {
+  status: string;
   e5: number;
   e10: number;
+  diesel: number;
+  lat: number;
+  lng: number;
+  dist: number;
+  name: string;
+  street: string;
   isOpen: boolean;
   houseNumber: string;
   postCode: number;
+  id: string;
 }
+
+
+
+
